added auth middleware and policy to projectctrl

+++ app/Http/Controllers/ProjectController.php
@@ -5,11 +5,19 @@ namespace App\Http\Controllers;
 use App\Models\project;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Support\Arr;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-
-class ProjectController extends Controller
+class ProjectController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -32,15 +40,13 @@ class ProjectController extends Controller
             'tags' => ['nullable']
         ]);
 
-        $project = Project::create(Arr::except($newProject, ['tags']));
+        $project = $request->user()->projects()->create(Arr::except($newProject, 'tags'));
 
-        // $project = $request->user()->projects()->create(Arr::except($newProject, 'tags'));
-
-        // if ($newProject['tags']) {
-        //     foreach (explode(',', $newProject['tags']) as $tag) {
-        //         $project->tag($tag);
-        //     }
-        // }
+        if ($newProject['tags']) {
+            foreach (explode(',', $newProject['tags']) as $tag) {
+                $project->tag($tag);
+            }
+        }
 
 
         return ['project' => $project];
@@ -59,6 +65,7 @@ class ProjectController extends Controller
      */
     public function update(Request $request, project $project)
     {
+        Gate::authorize('modify', $project);
         $req = $request->validate([
             'title' => ['required', 'min:5', 'max:10'],
             'address' => ['required'],
@@ -76,6 +83,7 @@ class ProjectController extends Controller
      */
     public function destroy(project $project)
     {
+        Gate::authorize('modify', $project);
         $project->delete();
 
         return ['message' => 'Project Deleted'];
